requirements.txt
asgiref==3.7.2
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.7
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-filter==23.4
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-smart-selects==1.6.0
djangorestframework==3.14.0
drf-spectacular==0.28.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
celery
requests==2.32.3
sqlparse==0.4.4
redis==5.0.8